refactor: extract theme logic into a dedicated ThemeProvider
